move mock file readers into their own files
